refactor(server): naming improvements
