+++ src/server/boot/setup.ts
@@ -1,13 +1,13 @@
 import mkdirp from 'mkdirp';
 import * as Sentry from '@sentry/node';
 
-import appConfig from '../config/app-config';
+import serverConfig from '../../config/serverConfig';
 
 export default async function setup(): Promise<void> {
     try {
-        mkdirp.sync(appConfig.workingCachePath);
-        mkdirp.sync(appConfig.librariesPath);
-        mkdirp.sync(appConfig.temporaryStoragePath);
+        mkdirp.sync(serverConfig.workingCachePath);
+        mkdirp.sync(serverConfig.librariesPath);
+        mkdirp.sync(serverConfig.temporaryStoragePath);
     } catch (error) {
         Sentry.captureException(error);
         throw error;
