+++ src/server/httpServer.ts
@@ -1,12 +1,14 @@
-import Logger from '../helper/Logger';
-import server from '../';
-import setup from './setup';
+import http from 'http';
+
+import app from './boot/app';
+import Logger from './helpers/Logger';
+import setup from './boot/setup';
 
 const log = new Logger('boot');
 
 export default async () => {
     await setup();
-
+    const server = http.createServer(app);
     return server.listen(process.env.PORT || 0, () => {
         log.info(`server booted on port ${(server.address() as any).port}`);
     });
