--- src/server/index.ts
@@ -1,5 +1,0 @@
-import http from 'http';
-
-import app from './boot/app';
-
-export default http.createServer(app);
